+++ CHANGES.rst
@@ -3,6 +3,23 @@ Change log
 
 .. note:: **Upgrade notes**: after upgrading, run the ``arouteserver setup-templates`` command to sync the local templates with those distributed with the new version. More details on the `Upgrading <https://arouteserver.readthedocs.io/en/latest/INSTALLATION.html#upgrading>`__ section of the documentation.
 
+v1.4.0
+------
+
+- New: Docker image to easily build route-server configurations.
+
+  For more details, see the `docker directory <https://github.com/pierky/arouteserver/tree/master/docker>`_.
+
+- Improvement: change the default value of ``bgpq3_path`` to ``bgpq4``.
+
+  The ``bgpq4`` tool is now referenced as the default one in the ``bgpq3_path`` configuration line of arouteserver.yml.
+
+  **Please note**: operators who are using the tool and who left the ``bgpq3_path`` configuration line unset will now need to either explicitly configure that line to point to their ``bgpq3`` binary or to make sure ``bgpq4`` is available on their system.
+
+- Fix: the ``ixf-member-export`` command now produces a JSON file compliant with `version 1.0 <https://github.com/euro-ix/json-schemas/blob/master/versions/ixp-member-list-1.0.schema.json>`_ of the `Euro-IX schema <https://github.com/euro-ix/json-schemas>`_.
+
+  See also `GitHub PR65 <https://github.com/pierky/arouteserver/pull/65>`_.
+
 v1.3.0
 ------